+++ src/mod/applications/mod_redis/mod_redis.c
@@ -260,7 +260,7 @@ SWITCH_LIMIT_RESET(limit_reset_redis)
 	
 		if ((keyc = credis_keys(redis, rediskey, uuids, switch_arraylen(uuids))) > 0) {
 			int i = 0;
-			int hostnamelen = strlen(switch_core_get_switchname())+1;
+			int hostnamelen = (int)strlen(switch_core_get_switchname())+1;
 			
 			for (i = 0; i < keyc && uuids[i]; i++){
 				const char *key = uuids[i] + hostnamelen;
